+++ VirtualTourist/Config.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2016 Max Saienko. All rights reserved.
 //
 
-import Foundation
-
 struct Config {
     struct API {
         static let Key = "9410986b8caa0a6c8c71e64f4b795ceb"
